Use custom submit button and move components to form folder
